TERMINADO

+++ ft_printf.h
@@ -6,18 +6,19 @@
 /*   By: vabad-ro <vabad-ro@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/01/28 15:32:38 by vabad-ro          #+#    #+#             */
-/*   Updated: 2026/01/29 21:15:03 by vabad-ro         ###   ########.fr       */
+/*   Updated: 2026/01/30 12:23:08 by vabad-ro         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef FT_PRINTF_H
 # define FT_PRINTF_H
 
-# include <stdio.h>
+# include <limits.h>
 # include <stdarg.h>
+# include <stdint.h>
+# include <stdio.h>
 # include <stdlib.h>
 # include <unistd.h>
-# include <stdint.h>
 
 int	ft_printf(char const *str, ...);
 int	ft_printchr(int c);
@@ -25,8 +26,7 @@ int	ft_printstr(char *s);
 int	ft_printnbr(int n);
 int	ft_printunbr(unsigned int n);
 int	ft_printptr(void *c);
-int	ft_printhexa(unsigned int c);
-int	ft_printhexamin(unsigned int c);
-
+int	ft_printhexa(unsigned long c);
+int	ft_printhexamin(unsigned long c);
 
 #endif
